refactor(utils): move unicode17 out of the ecmascript package
Nothing in the package is about ECMAScript: it says which characters Unicode
16 and 17 gave a case to, and which of them the two properties behind
Final_Sigma took on. Thirty packages here read the standard library's unicode
tables and about twenty of those call sites ask a character whether it is
upper or lower case, so the gap it stands in is the whole repository's, not
one package's.
